Fileretrieval fixes and storage rename
