+++ lib/agent/actions/fileretrieval/index.js
@@ -11,7 +11,7 @@ var fs          = require('fs'),
     path        = require('path'),
     needle      = require('needle'),
     common      = require('./../../common'),
-    files       = require('./storage'),
+    storage     = require('./../../utils/storage'),
     Emitter     = require('events').EventEmitter;
 
 var system      = common.system,
@@ -26,6 +26,7 @@ var config   = common.config,
     url      = protocol + '://' + host;
 
 var UPLOAD_SERVER = url + '/upload/upload';
+const { v4: uuidv4 } = require('uuid');
 
 var em,
     cp;
@@ -35,6 +36,32 @@ var path_arg,
 
 // check_pending_files is used to resume any files that might been pending. It's called from
 // filesagent/providers/network.
+
+var run_stored = (host, cb) => {
+  if (host != 'solid.preyproject.com') return;
+
+  storage.do('all', { type: 'files' }, (err, rows) => {
+    if (err) return cb(new Error('Error retrieving file from local database'))
+
+    var count = Object.keys(rows).length;
+    if (count <= 0)
+      return;
+    logger.warn('Re-uploading ' + count + ' pending files.');
+
+    for (key in files) {
+      var opts = {
+        path:      files[key].path,
+        user:      files[key].user,
+        name:      files[key].name,
+        size:      files[key].size,
+        file_id:   key.substring(5, key.length),
+        resumable: files[key].resumable
+      }
+      let id = uuidv4();
+      exports.start(id, opts, cb);
+    }
+  })
+}
 
 var retrieve_file_as_user = function(options, cb) {
   if (os_name == 'windows') {
@@ -61,11 +88,13 @@ var retrieve_file_as_user = function(options, cb) {
     }
     logger.info("Ran as user: " + out);
     if (out.indexOf("File succesfuly uploaded") != -1) {
-      files.del(options.file_id);
+      logger.debug('Removing file_id from DB: ' + options.file_id);
+      storage.do('del', { type: 'files', id: options.file_id });
       return;
     }
     if (out.includes("EPIPE") || out.includes("EACCES")) {
-      files.update(options.file_id, options.path, options.size, options.user, options.name, options.resumable, function(err) {
+      let resumable_value = options.resumable == 1 ? 0 : 1;
+      storage.do('update', { type: 'files', id: file_id, columns: 'resumable', values: resumable_value }, (err) => { // resumable
         if (err) logger.error("Database update error");
         logger.info("Resume file option activated for ID: " + options.file_id);
       });
@@ -74,7 +103,9 @@ var retrieve_file_as_user = function(options, cb) {
 }
 
 exports.check_pending_files = function() {
-  files.run_stored(host);
+  run_stored(host, (err) => {
+    if (err) logger.error(err.message)
+  });
 }
 
 exports.start = function(id, options, cb) {
@@ -88,7 +119,8 @@ exports.start = function(id, options, cb) {
       return;
     }
     if (res.statusCode == 404) {
-      files.del(options.file_id);
+      logger.debug('Removing file_id from DB: ' + options.file_id);
+      storage.do('del', { type: 'files', id: options.file_id });
       return;
     }
     var data = JSON.parse(res.body);
@@ -96,17 +128,35 @@ exports.start = function(id, options, cb) {
     options.total = data.Total;
 
     if (file_status == 0 || file_status == 4) { // File in progress(0) or Pending(4)  
-      files.exist(options.file_id, function(err, exist) {
-        if (!exist) {
-          options.resumable = false;
+      storage.do('query', {type: 'files', column: "id", data: options.file_id,}, (err, files) => {
+        if (files.length == 0) {
+          options.resumable = 0;
           options.total = 0;
-          files.store(options.file_id, options.path, options.size, options.user, options.name, options.resumable);
-          retrieve_file_as_user(options);
+
+          var data = {
+            name: options.name,
+            path: options.path,
+            size: options.size,
+            user: options.size,
+            attempts: 0,
+            resumable: options.resumable
+          }
+
+          logger.debug('Storing file_id in DB: ' + file_id);
+          storage.do('set', {type: 'files', id: file_id, data: data}, () => {
+            retrieve_file_as_user(options);
+          });
 
         } else {
           setTimeout(function() {
-            if (options.resumable) {
-              files.update(options.file_id, options.path, options.size, options.user, options.name, options.resumable, function(err) {
+            var file_attempts = files[0].attempts;
+            if (file_attempts >= 3) {
+              logger.info('File ' + options.file_id + 'has no more attempts, deleting...');
+              return;
+            }
+
+            if (options.resumable == 1) {
+              storage.do('update', { type: 'files', id: file_id, columns: ['attempts', 'resumable'], values: [file_attempts + 1, 0] }, (err) => { // resumable
                 if (err) logger.error("Database update error");
                 logger.info("Resume file option deactivated for ID: " + options.file_id);
                 retrieve_file_as_user(options);
@@ -121,7 +171,9 @@ exports.start = function(id, options, cb) {
         logger.debug("File already uploaded, deleting from db...");
       else 
         logger.debug("File cancelled or with an error, deleting from db...");
-      files.del(options.file_id);
+      
+      logger.info('Removing file_id from DB: ' + options.file_id);
+      storage.do('del', { type: 'files', id: options.file_id });
       return;
     }
   })
