+++ lib/agent/actions/geofencing/index.js
@@ -3,9 +3,9 @@
 var async        = require('async'),
     EventEmitter = require('events').EventEmitter,
     api          = require('./../../plugins/control-panel/api'),
+    storage      = require('./../../utils/storage'),
     hooks        = require('./../../hooks');
 
-var new_storage = require('./../../utils/commands_storage');
 var emitter;
 
 exports.watching;
@@ -15,7 +15,7 @@ function fetch_geofences(cb) {
 }
 
 exports.get_geofences = function (cb) {
-  new_storage.do('all', { type: 'geofences' }, (err, stored_geofences) => {
+  storage.do('all', { type: 'geofences' }, (err, stored_geofences) => {
     if (err) return cb(new Error ('Error retrieving geonfences from local database'))
     return cb(null, stored_geofences);
   })
@@ -27,7 +27,7 @@ exports.sync = function(id, geofences, cb) {
   let fences = geofences;
 
   // Se obtienen las zonas locales almacenadas
-  new_storage.do('all', { type: 'geofences' }, (err, stored_geofences) => {
+  storage.do('all', { type: 'geofences' }, (err, stored_geofences) => {
     let array = [];
 
     // Por cada stored fence se revisa si existe en el listado nuevo
@@ -47,7 +47,7 @@ exports.sync = function(id, geofences, cb) {
           //Si no existe en el listado nuevo se borra del local, se quita del listado de nuevas...
         } else {
           // se quita de las zonas locales
-          new_storage.do('del', {type: 'geofences', id: stored_fence.id}, callback);
+          storage.do('del', {type: 'geofences', id: stored_fence.id}, callback);
         }
       })
     })
@@ -59,7 +59,7 @@ exports.sync = function(id, geofences, cb) {
       fences.forEach(fence => {
         array2.push((callback) => {
           exports.watching.push(parseInt(fence.id));
-          new_storage.do('set', {type: 'geofences', id: fence.id, data: {name: fence.name, state: 'NULL'}}, callback)
+          storage.do('set', {type: 'geofences', id: fence.id, data: {name: fence.name, state: 'NULL'}}, callback)
         }) 
       })
