+++ lib/agent/actions/triggers/index.js
@@ -6,9 +6,8 @@ var schedule     = require('node-schedule'),
     api          = require('./../../plugins/control-panel/api'),
     hooks        = require('./../../hooks'),
     commands     = require('./../../commands'),
+    storage      = require('./../../utils/storage'),
     lp           = require('./../../plugins/control-panel/long-polling');
-
-var storage2 = require('./../../utils/commands_storage');
 
 var emitter,
     running_triggers = [],
@@ -110,10 +109,10 @@ function check_rules(rule) {
 
 function run_trigger_actions(trigger) {
   // Update last_exec
-  storage2.do('update', {type: 'triggers', id: trigger.id, columns: 'last_exec', values: new Date().getTime() }, () => {
+  storage.do('update', {type: 'triggers', id: trigger.id, columns: 'last_exec', values: new Date().getTime() }, () => {
       // send_response('stopped', [trigger.id]);  ---> revisar!!!
       // storage.del(trigger.id);
-      storage2.do('del', {type: 'triggers', id: trigger.id});
+      storage.do('del', {type: 'triggers', id: trigger.id});
       return;
   })
 
@@ -212,7 +211,7 @@ function set_up_hooks() {
           da_triggers[trigger_index].last_exec = exec_time;
 
           var current = da_triggers[trigger_index];
-          storage2.do('update', {type: 'triggers', id: current.id, columns: 'synced_at', values: exec_time}, (err) => {
+          storage.do('update', {type: 'triggers', id: current.id, columns: 'synced_at', values: exec_time}, (err) => {
             if (err) logger.info("Unable to update the execution time");
           });
         }
@@ -502,8 +501,7 @@ exports.sync = (id, err, triggers, stored, cb) => {
       if (trigger.persist && (trigger.persist == true || trigger.persist == 1))
         data.persist = 1;
 
-
-      storage2.do('set', {type: 'triggers', id: trigger.id, data: data}, (err) => {
+      storage.do('set', {type: 'triggers', id: trigger.id, data: data}, (err) => {
         if (err) logger.error("Error storing triggers: " + err);
 
         if (index == triggers.length - 1) {
@@ -535,7 +533,7 @@ function refresh_triggers(id, opts, cb) {
 
   fetch_triggers((err, res) => {
     if (err) {
-      storage2.do('all', {type: 'triggers'}, (err, stored_triggers) => {
+      storage.do('all', {type: 'triggers'}, (err, stored_triggers) => {
         if (err) return done(id, err, cb);
         exports.clear_triggers((err) => {
           exports.sync(id, err, [], stored_triggers, () => {
@@ -552,7 +550,7 @@ function refresh_triggers(id, opts, cb) {
         return done(id, err, cb);
       }
 
-      storage2.do('all', {type: 'triggers'}, (err, stored_triggers) => {
+      storage.do('all', {type: 'triggers'}, (err, stored_triggers) => {
         exports.clear_triggers((err) => {
           exports.sync(id, null, triggers, stored_triggers, () => {
             cb && cb(null, emitter);
@@ -564,7 +562,7 @@ function refresh_triggers(id, opts, cb) {
 }
 
 exports.clear_triggers = (cb) => {
-  storage2.do('clear', {type: 'triggers'}, (err) => {
+  storage.do('clear', {type: 'triggers'}, (err) => {
     if (err) logger.error(err.message);
     return cb() && cb(err);
   });
