+++ lib/agent/commands.js
@@ -5,7 +5,7 @@ var common    = require('./common'),
     providers = require('./providers'),
     reports   = require('./reports'),
     updater   = require('./updater'),
-    new_storage = require('./utils/commands_storage'),
+    storage = require('./utils/storage'),
     devices   = require('./plugins/control-panel/api/devices');
 
 const { v4: uuidv4 } = require('uuid');
@@ -217,11 +217,11 @@ exports.store = store;
 
 // When storing an action it gets deleted only when an action with the same name arrives.
 var delete_same_target = (id, target, cb) => {
-  new_storage.do('query', {type: 'commands', column: "target", data: target}, (err, actions) => {
+  storage.do('query', {type: 'commands', column: "target", data: target}, (err, actions) => {
     if (actions.length == 0) return cb();
     actions.forEach((action, index) => {
       if (id != action.id) {
-        new_storage.do('del', {type: 'commands', id: action.id}, () => {
+        storage.do('del', {type: 'commands', id: action.id}, () => {
           if (index == actions.length - 1) return cb();
         });
       }
@@ -235,17 +235,17 @@ var store = function(type, id, name, opts, cb) {
 
   delete_same_target(id, name, () => {
     if (name == 'geofencing' || name == 'triggers' || name == 'fileretrieval') return cb && cb();
-    new_storage.do('set', {type: 'commands', id: id, data: {command: type, target: name, options: opts }}, cb);
+    storage.do('set', {type: 'commands', id: id, data: {command: type, target: name, options: opts }}, cb);
   })
 }
 
 var remove = function(type, id, name, cb) {
   logger.debug('Removing command from DB: ' + [type, name].join('-'));
-  new_storage.do('del', { type: 'commands', id: id }, cb)
+  storage.do('del', { type: 'commands', id: id }, cb)
 }
 
 var verify_if_executed = function(id, cb) {
-  new_storage.do('query', {type: 'commands', column: "id", data: id}, function (err,rows) {
+  storage.do('query', {type: 'commands', column: "id", data: id}, function (err,rows) {
     if (err) return cb(err);
 
     if (rows.length == 0 || (rows[0] && rows[0].stopped == 'NULL')) return cb(null, false);
@@ -273,13 +273,13 @@ var watch_stopped = function() {
 
   hooks.on('action', function(event, id) { 
     if (event == 'stopped' || event == 'failed') {
-      new_storage.do('update', { type: 'commands', id: id, columns: 'stopped', values: new Date().toISOString() }, (err) => {
+      storage.do('update', { type: 'commands', id: id, columns: 'stopped', values: new Date().toISOString() }, (err) => {
         if (err) logger.warn("Unable to update stopped action timestamp for id:" + id)
       })
     }
 
     if (event == 'started') {
-      new_storage.do('update', { type: 'commands', id: id, columns: 'started', values: new Date().toISOString() }, (err) => {
+      storage.do('update', { type: 'commands', id: id, columns: 'started', values: new Date().toISOString() }, (err) => {
         if (err) logger.warn("Unable to update started action timestamp for id:" + id)
       })
     }
@@ -301,7 +301,7 @@ exports.stop_watching = function() {
 }
 
 exports.run_stored = function(cb) {
-  new_storage.do('all', {type: 'commands'}, (err, commands) => {
+  storage.do('all', {type: 'commands'}, (err, commands) => {
     if (err)
       return logger.error(err.message);
