+++ lib/agent/plugins/control-panel/long-polling/index.js
@@ -9,7 +9,7 @@ var needle        = require('needle'),
     triggers      = require('./../../../actions/triggers'),
     lp_conf       = require('./../../../../conf/long-polling'),
     status_trigger = require('./../../../triggers/status'),
-    storage       = require('./../../../utils/commands_storage'),
+    storage       = require('./../../../utils/storage'),
     updater       = require('./../../../updater'),
     server        = require('./server'),
     Emitter       = require('events').EventEmitter;
