+++ lib/agent/plugins/control-panel/long-polling/test/index_spec.js
@@ -11,7 +11,7 @@ var assert  = require('assert'),
     common  = require('../../../../../common'),
     status  = require('./../../../../triggers/status'),
     EventEmitter = require('events').EventEmitter,
-    storage   = require('./../../../../utils/commands_storage'),
+    storage   = require('./../../../../utils/storage'),
     status_trigger = require('./../../../../triggers/status'),
     server       = require('./../server'),
     test_server  = require('./test_server');
