+++ lib/agent/plugins/control-panel/secure.js
@@ -3,15 +3,13 @@ var JSEncrypt   = require('node-jsencrypt'),
     keys        = require('./api/keys'),
     devices     = require('./api/devices'),
     common      = require('./../../common'),
-    new_storage  = require('./../../utils/commands_storage'),
+    storage     = require('./../../utils/storage'),
     system      = require('./../../../system'),
     account     = require('./../../../conf/account'),
     exec        = cp.exec,
     run_as_user = system.run_as_logged_user,
     logger      = common.logger.prefix('config'),
     os_name     = process.platform.replace('darwin', 'mac').replace('win32', 'windows');
-
-// new_storage.init('keys');
 
 var crypt,
     data,
@@ -69,7 +67,7 @@ var format_public_key = function(public_key, cb) {
 }
 
 exports.generate_keys = function(cb) {
-  new_storage.do('all', {type: 'keys'}, (err, values) => {
+  storage.do('all', {type: 'keys'}, (err, values) => {
     if (err) return cb(new Error("Error reading stored security keys"));
 
     // Read stored keys if they exists
@@ -95,9 +93,9 @@ exports.generate_keys = function(cb) {
       crypt.setPublicKey(public_key);
       crypt.setPrivateKey(private_key);
 
-      new_storage.do('set', {type: 'keys', id: 'public_key', data: {value: Buffer.from(public_key).toString('base64')}}, (err) => {
+      storage.do('set', {type: 'keys', id: 'public_key', data: {value: Buffer.from(public_key).toString('base64')}}, (err) => {
         if (err) return cb(new Error("Error storing public security key"));
-        new_storage.do('set', {type: 'keys', id: 'private_key', data: {value: Buffer.from(private_key).toString('base64')}}, (err) => {
+        storage.do('set', {type: 'keys', id: 'private_key', data: {value: Buffer.from(private_key).toString('base64')}}, (err) => {
           if (err) return cb(new Error("Error storing private security key"));
           format_public_key(public_key, cb);
         })
@@ -141,9 +139,9 @@ exports.open_config = function(device_key, cb) {
 }
 
 exports.reset_keys = function(cb) {
-  new_storage.do('del', {type: 'keys', id: 'private_key'}, (err) => {
+  storage.do('del', {type: 'keys', id: 'private_key'}, (err) => {
     if (err) return cb(err);
-    new_storage.do('del', {type: 'keys', id: 'private_key'}, (err) => {
+    storage.do('del', {type: 'keys', id: 'private_key'}, (err) => {
       if (err) return cb(err);
       public_keys = {};
       exports.generate_keys(cb);
