+++ lib/agent/plugins/control-panel/test/secure.js
@@ -4,7 +4,7 @@ var should      = require('should'),
     secure      = require('../secure'),
     devices     = require('./../api/devices');
     account     = require('./../../../../conf/account'),
-    storage2    = require('./../../../utils/commands_storage'),
+    storage     = require('./../../../utils/storage'),
     tmpdir      = require('os').tmpdir;
 
 var file,
@@ -24,9 +24,9 @@ describe('secure()', function() {
 
     describe('with keys stored', function() {
       before(function(done) {
-        storage2.init('keys', tmpdir() + '/store2.db', () => {
-          storage2.do('set', {type: 'keys', id: 'public_key', data: {value: Buffer.from(public_key).toString('base64') }} , function(err) {
-            storage2.do('set', {type: 'keys', id: 'private_key', data: {value: Buffer.from(private_key).toString('base64') }} , function(err) {
+        storage.init('keys', tmpdir() + '/store2.db', () => {
+          storage.do('set', {type: 'keys', id: 'public_key', data: {value: Buffer.from(public_key).toString('base64') }} , function(err) {
+            storage.do('set', {type: 'keys', id: 'private_key', data: {value: Buffer.from(private_key).toString('base64') }} , function(err) {
             done();
             })
           })
@@ -34,7 +34,7 @@ describe('secure()', function() {
       })
 
       after((done) => {
-        storage2.erase(tmpdir() + '/store2.db', done);
+        storage.erase(tmpdir() + '/store2.db', done);
       })
 
       it('returns the stored keys', function(done) {
@@ -49,11 +49,11 @@ describe('secure()', function() {
 
     describe('with no keys stored', function() {
       before(function(done) {
-        storage2.init('keys', tmpdir() + '/store2.db', done)
+        storage.init('keys', tmpdir() + '/store2.db', done)
       })
 
       after((done) => {
-        storage2.erase(tmpdir() + '/store2.db', done);
+        storage.erase(tmpdir() + '/store2.db', done);
       })
         
       it('generate, store and returns new keys', function(done) {
@@ -62,7 +62,7 @@ describe('secure()', function() {
           publicKey = secure.public_keys().default;
           publicKey.should.containEql('BEGIN PUBLIC KEY-----');
           publicKey.should.not.be.equal(public_key);
-          storage2.do('query', {type: 'keys', column : 'id', data: 'public_key' }, function (err,rows) {
+          storage.do('query', {type: 'keys', column : 'id', data: 'public_key' }, function (err,rows) {
             let key = rows[0];
             should.exist(key);
             key.value.should.equal(Buffer.from(publicKey).toString('base64'));
@@ -115,7 +115,7 @@ describe('secure()', function() {
           return cb(null);
         })
 
-        storage2.init('keys', tmpdir() + '/store2.db', () => {
+        storage.init('keys', tmpdir() + '/store2.db', () => {
           secure.generate_keys(function() {
             crypt.setPublicKey(secure.public_keys().default);
             encrypted_key = crypt.encrypt(api_key);
@@ -126,7 +126,7 @@ describe('secure()', function() {
 
       after(function(done) {
         auth_stub.restore();
-        storage2.erase(tmpdir() + '/store2.db', done);
+        storage.erase(tmpdir() + '/store2.db', done);
       })
 
       it('doesnt returns error', function(done) {
