+++ lib/agent/providers/hardware/index.js
@@ -15,7 +15,7 @@ var _       = require('underscore'),
     common  = require('./../../common'),
     logger  = common.logger.prefix('hardware'),
     os_name = process.platform.replace('darwin', 'mac').replace('win32', 'windows'),
-    storage2 = require('./../../utils/commands_storage'),
+    storage = require('./../../utils/storage'),
     os_functions = require('./' + os_name),
     exp     = module.exports;
 
@@ -84,7 +84,7 @@ exp.track_hardware_changes = (data) => {
 
   var diff_count = 0;
   var save_data = () => {
-    storage2.do('set', {type: 'keys', id: 'hardware', data: {value: JSON.stringify(data)}}, (err) => {
+    storage.do('set', {type: 'keys', id: 'hardware', data: {value: JSON.stringify(data)}}, (err) => {
       if (err) logger.error('Unable to save hardware data');
     });
   }
@@ -107,7 +107,7 @@ exp.track_hardware_changes = (data) => {
     }
   }
 
-  storage2.do('query', {type: 'keys', column: 'id', data: 'hardware'}, (err, stored_data) => {
+  storage.do('query', {type: 'keys', column: 'id', data: 'hardware'}, (err, stored_data) => {
     if (err) logger.error('Unable to read hardware data');
     if (stored_data.length == 0) return save_data();
 
@@ -120,7 +120,7 @@ exp.track_hardware_changes = (data) => {
 
     if (diff_count > 0) {
       hooks.trigger('hardware_changed');
-      storage2.do('del', {type: 'keys', id: 'hardware'}, (err) => {
+      storage.do('del', {type: 'keys', id: 'hardware'}, (err) => {
         if (err) logger.error('Unable to delete hardware data');
         save_data();
       })
