+++ lib/agent/triggers/control-zones/index.js
@@ -32,8 +32,7 @@ var join      = require('path').join,
     logger    = require('./../../common').logger.prefix('geofencing'),
     LatLon    = require('./../location/lib/latlng'),
     action    = require('./../../actions/geofencing'),
-    // storage   = require('./../../actions/geofencing/storage'),
-    new_storage  = require('./../../utils/commands_storage'),
+    storage   = require('./../../utils/storage'),
     api       = require('./../../plugins/control-panel/api'),
     Emitter   = require('events').EventEmitter;
 
@@ -92,7 +91,7 @@ var check_zones = function() {
 
     if (distance * 1000 > zone.radius) { // outside
 
-      new_storage.do('update', { type: 'geofences', id: zone.id, columns: 'state', values: 'outside' }, (err) => {
+      storage.do('update', { type: 'geofences', id: zone.id, columns: 'state', values: 'outside' }, (err) => {
           if (err) logger.error(err);
           if (zone.state == 'inside' && zone.direction !== 'in') {
             logger.info('Device left the geofence ' + zone.name + '! Notifying')
@@ -105,7 +104,7 @@ var check_zones = function() {
   
     } else { // inside
 
-      new_storage.do('update', { type: 'geofences', id: zone.id, columns: 'state', values: 'inside' }, (err) => {
+      storage.do('update', { type: 'geofences', id: zone.id, columns: 'state', values: 'inside' }, (err) => {
         if (err) logger.error(err);
         if ((zone.state == 'outside' && zone.direction !== 'out') || zone.state == null) {
           logger.info('Device got inside the geofence ' + zone.name + '! Notifying')
