+++ lib/agent/triggers/hostname/index.js
@@ -4,7 +4,7 @@ var triggers    = require('os-triggers'),
     path        = require('path'),
     Emitter     = require('events').EventEmitter,
     common      = require('./../../common'),
-    storage     = require('./../../utils/commands_storage'),
+    storage     = require('./../../utils/storage'),
     base_path   = path.join(__dirname, '..', '..'),
     hooks       = require(path.join(base_path, 'hooks')),
     providers   = require(path.join(base_path, 'providers')),
