+++ lib/agent/utils/storage.js
@@ -1,237 +1,330 @@
-var fs       = require('fs'),
-    join     = require('path').join,
-    sqlite3  = require('sqlite3').verbose(),
-    config   = require('./../../system/paths').config;
-    // fileType = require('./../providers/file-type');
+var fs         = require('fs'),
+    async      = require('async'),
+    join       = require('path').join,
+    sqlite3    = require('sqlite3').verbose(),
+    config     = require('./../../system/paths').config;
 
-var storage_path, db_type, db_comm;
+const { v4: uuidv4 } = require('uuid');
+
+var storage_path,
+    backup_path,
+    db_comm,
+    first_init = true;
+
+var singular = function(type) {
+  return type.substring(0, type.length - 1);
+}
+
+var types = {
+  commands: {
+    schema: 'id TEXT PRIMARY KEY, command TEXT, target TEXT, options TEXT, started_resp INTEGER, started INTEGER, stopped_resp INTEGER, stopped INTEGER',
+    keys: ['id', 'command', 'target', 'options', 'started_resp', 'started', 'stopped_resp', 'stopped'],
+    values: (data) => { return `'${data.id}', '${data.command}', '${data.target}', '${JSON.stringify(data.options)}' , 0, 'NULL', 0, 'NULL'` }
+  },
+  geofences: {
+    schema: 'id TEXT PRIMARY KEY, name TEXT, state TEXT',
+    keys: ['id', 'name', 'state'],
+    values: (data) => { return `'${data.id}', '${data.name}', '${data.state}'` }
+  },
+  files: {
+    schema: 'id TEXT PRIMARY KEY, name TEXT, path TEXT, size TEXT, user TEXT, attempt INTEGER, resumable INTEGER',
+    keys: ['id', 'name', 'path', 'size', 'user', 'attempt', 'resumable'],
+    values: (data) => { return `'${data.id}', '${data.name}', '${data.path}', '${data.size}', '${data.user}', 0, 0` }
+  },
+  triggers: {
+    schema: 'id TEXT PRIMARY KEY, name TEXT, persist INTEGER, synced_at INTEGER, last_exec INTEGER, automation_events TEXT, automation_actions TEXT',
+    keys: ['id', 'name', 'persist', 'synced_at', 'last_exec', 'automation_events', 'automation_actions'],
+    values: (data) => { return `'${data.id}', '${data.name}', '${data.persist}', '${data.synced_at}', '${data.last_exec}', '${JSON.stringify(data.automation_events)}', '${JSON.stringify(data.automation_actions)}'` }
+  },
+  keys: {
+    schema: 'id TEXT PRIMARY KEY, value TEXT',
+    keys: ['id', 'value'],
+    values: (data) => { return `'${data.id}', '${data.value}'`}
+  },
+  versions: {
+    schema: 'id TEXT PRIMARY KEY, "from" TEXT, "to" TEXT, attempts INTEGER, notified INTEGER',
+    keys: ['id', '"from"', '"to"', 'attempts', 'notified'],
+    values: (data) => { return `'${data.id}', '${data.from}', '${data.to}','${data.attempts}','${data.notified}'`}
+  }
+}
 
 var SQLITE_ACCESS_ERR = "Access denied to commands database, must run agent as prey user"
 exports.store_path = null;
 
-var queries = {
-  CREATE: function(type)      { return "CREATE TABLE IF NOT EXISTS " + type + " (" + singular(type) + " text)" },
-  INSERT: function(type)      { return "INSERT INTO " + type + " VALUES (?)" },
-  SELECT: function(type)      { return "SELECT * FROM " + type },
-  DELETE: function(type)      { return "DELETE FROM " + type },
-  DROP:   function(type, del) { return type + " WHERE " + singular(type) + " = " + '"'+ del +'"' }
+var queries = {                                                                  
+  CREATE: (table)              => { 
+    // console.log("!!!",`CREATE TABLE IF NOT EXISTS ${table} (${types[table].schema})` )
+    return `CREATE TABLE IF NOT EXISTS ${table} (${types[table].schema})` },
+  INSERT: (table, data)        => { return `INSERT INTO ${table} (${types[table].keys.join(', ')}) VALUES (${types[table].values(data)})` },
+  DELETE: (table, id)          => { return `DELETE FROM ${table} WHERE id = '${id}'` },
+  CLEAR:  (table)              => { return `DELETE FROM ${table}` },
+  DROP:   (table)              => { return `DROP TABLE IF EXISTS ${table}`},
+  UPDATE: (table, id, changes) => { return `UPDATE ${table} SET ${changes} WHERE id = '${id}'` },
+  SELECT: (table)              => { return `SELECT * FROM ${table}` },
+  QUERY:  (table, query)       => { return `SELECT * FROM ${table} WHERE ${query}`},
+  INFO:   (table)              => { return `SELECT * FROM PRAGMA_TABLE_INFO('${table}')`},
+  TABLES: ()                   => { return `SELECT name FROM sqlite_master`}
 }
 
-var singular = function(db_type) {
-  return db_type.substring(0, db_type.length - 1);
+var validate_type = (type) => {
+  let valid = Object.keys(types).indexOf(type) == -1 ? false : true;
+  return valid;
 }
 
-var check_type = function(key, cb) {
-  if (key.includes('start') || key.includes('report-')) return cb(null, 'commands');
-  else if (key.includes('file-'))                       return cb(null, 'files');
-  else if (key.includes('geofence'))                    return cb(null, 'geofences');
-  else if (key.includes('trigger'))                     return cb(null, 'triggers');
-  else if (key.includes('version'))                     return cb(null, 'versions');
-  else if (key.includes('key'))                         return cb(null, 'keys');
-  else return cb(new Error("Not an allowed type of key"))
+var storage_queue = async.queue((task, cb) => {
+  storage_fns[task.perform](task.opts, cb);
+}, 1);
+
+// All storage interaction goes through here...
+exports.do = (perform, opts, cb) => {
+  // Pause queue current execution
+  storage_queue.pause();
+
+  // Queue the new command
+  storage_queue.push({perform: perform, opts: opts}, cb);
+
+  // Resume queue
+  storage_queue.resume();
 }
 
-var load = function(type, cb) {
-  var db;
-  if (!storage_path) {
-    return cb(new Error('Invalid path'), null);
-  }
-
-  db_comm.all(queries.SELECT(type), function(err, rows) {
-    if (err) {
-      if (err.code != 'ENOENT')
-        return cb(err, null);
-
-      db = {};
-      return cb(err, db);
-    }
-
-    try {
-      var db1 = {};
-      for (var i = 0; i < rows.length; i++) {
-        var value = JSON.parse(Buffer.from(rows[i][singular(type)], 'base64').toString());
-        var key = Object.keys(value)[0];
-        db1[key] = value[key];
-      }
-      db = db1;
-    } catch(e) {
-      db = {};
-    }
-
-    cb(null, db);
-  })
-}
-
-var remove = function(type, cb) {
-  if (!storage_path) {
-    // return cb(new Error('Invalid path'));
-    return cb();
-  }
-
-  db_comm.run(queries.DELETE(type), function(err) {
-    var e = err && err.code != 'ENOENT' ? err : null;
-    return cb && cb(e);
-  })
-}
-
-var save = function(type, add, cb) {
-  var key = Object.keys(add)[0];
-  exports.del(key, function(err) {
-    var stmt   = db_comm.prepare(queries.INSERT(type));
-    var to_add = Buffer.from(JSON.stringify(add, null, 0)).toString('base64');
-
-    stmt.run(to_add, function(err) {
-      var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
-      stmt.finalize();
-      return cb && cb(e);
-    });
-  })
-}
-
-var drop = function(type, del, cb) {
-  var to_del   = Buffer.from(JSON.stringify(del, null, 0)).toString('base64');
-  var del_comm = queries.DROP(type, to_del);
-
-  db_comm.all(queries.DELETE(del_comm), function(err) {
-    var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
-    return cb && cb(e);
-  });
-}
-
-var check_comm_type = function(type, path, cb) {
-  if (type == 'commands') {
-    fileType.get_file_type(path, function(err, out){
-      if (err) return cb();
-
-      if (out == "ASCII") {
-        fs.unlink(path, function(err) {
-          var e = err && err.code != 'ENOENT' ? err : null;
-          return cb();
-        });
-      } else {
-        return cb();
-      }
-    });
-  } else {
-    cb();
-  }
-}
-
-var check_and_init = function(key, cb) {
-  check_type(key, function(err, type) {
-    if (err) return cb && cb(err);
-    init(type, '', function(err) {
-      if (err) return cb && cb(err);
-      
-      cb(null, type)
-    });
-  });
-}
+// Queue's callback...
+storage_queue.drain = function() {
+  console.log('Queue is Done!');
+};
 
 var init = function(type, path, cb) {
   storage_path = path ? path : (storage_path ? storage_path : join(config, 'commands.db'));
-  exports.store_path = storage_path;
+  backup_path = storage_path;
 
-  // check_comm_type(type, storage_path, function() {
+  var create_db = () => {
     db_comm = new sqlite3.Database(storage_path, function(err) {
       if (err) {
         if (err.code == 'SQLITE_CANTOPEN')
           return cb(new Error(SQLITE_ACCESS_ERR))
         return cb(err)
       }
-      db_comm.run(queries.CREATE(type), function(err) {
-        return cb && cb(err);
-      });
+
+      if (type) {
+        if (!validate_type(type)) return cb(new Error('Not an allowed type of key'))
+        db_comm.run(queries.CREATE(type), function(err) {
+          return cb && cb(err, db_comm);
+        });
+      } else return cb(null, db_comm);
+      // } else return cb(new Error('Must declare a valid type of table'));
     });
-  // });
-}
+  }
 
-exports.set = function(key, data, cb) {
-  check_and_init(key, function(err, type) {
-    if (err) return cb && cb(err);
-
-    var to_add = {};
-    to_add[key] = data;
-    save(type, to_add, cb);
-  });
-}
-
-exports.get = function(key, cb) {
-  check_type(key, function(err, type) {
-    if (err) return cb && cb(err);
-    load(type, function(err, db) {
-      if (err) return cb && cb(err);
-
-      cb(null, db[key]);
-    })
-  });
-}
-
-exports.del = function(key, cb) {
-  check_and_init(key, function(err, type) {
-    if (err) return cb && cb(err);
-    load(type, function(err, db) {
-      if (err) return cb && cb(err);
-
-      var found = false;
-      var to_del = {};
-
-      for (var k in db) {
-        if (k == key) {
-          found = true;
-          to_del[key] = db[k];
-        }
+  if (first_init) {
+    var existing_db = new sqlite3.Database(storage_path, function(err) {
+      if (err) {
+        erase(storage_path, () => {
+          storage_path = backup_path;
+          create_db();
+        });
+      } else {
+        recover_db(existing_db, (err) => {
+          if (err) erase(storage_path, () => {
+            storage_path = backup_path;
+          });
+          first_init = false;
+          return create_db();
+        })
       }
+    });
 
-      return found ? drop(type, to_del, cb) : cb && cb();
-    })
-  });
+  } else create_db();
 }
 
-exports.update = function(key, del, add, cb) {
-  check_and_init(key, function(err, type) {
-    if (err) return cb && cb(err);
+var storage_fns = {};
 
-    drop(type, del, function(err) {
-      if (err) return cb() && cb(err);
-      save(type, add, cb);
-    })
-  });
-}
+storage_fns.set = (opts, cb) => {
+  init(opts.type, null, (err) => {
+    if (opts.id) opts.data.id = opts.id
 
-exports.all = function(type, cb) {
-  init(type, '', function(err) {
-    if (err) return cb(err);
-    load(type, function(err, db) {
-      if (err) return cb(err);
+    let column = "id",
+        data = opts.data,
+        query = `${column} = '${data.id}'`;
 
-      cb && cb(null, db);
+   db_comm.all(queries.QUERY(opts.type, query), function(err, rows) {
+    if (err || !rows) {
+      if (err.code != 'ENOENT')
+        return cb(err, null);
+      return cb(err, []);
+    }
+    if (rows.length == 0) {
+
+      db_comm.run(queries.INSERT(opts.type, opts.data), (err) => {
+        var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
+        return cb && cb(e);
+      });
+    }
+    else return cb(new Error(`Already registered a ${singular(opts.type)} with id: ${opts.id}`))
     })
   })
 }
 
-// empties db and removes file
-exports.clear = function(type, cb) {
-  init(type, '', function(err) {
-    if (err) return cb(err);
-    remove(type, function(err) {
-      cb && cb(err);
+storage_fns.del = (opts, cb) => {
+  init(opts.type, null, (err) => {
+    db_comm.run(queries.DELETE(opts.type, opts.id), function(err) {
+      var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
+      return cb && cb(e);
+    })
+  })
+}
+
+storage_fns.update = (opts, cb) => {
+  init(opts.type, null, (err) => {
+    let changes,
+        values = opts.values,
+        columns = opts.columns;    
+
+    if (Array.isArray(columns)) {
+
+      let data = [];
+
+      if (columns.length != values.length || columns.length == 0)
+        return cb(new Error(''));
+
+      columns.forEach((el, index) => {
+        let change = `${el} = '${values[index]}'`
+        data.push(change)
+      });
+
+      changes = data.join(', '); // duda
+
+    } else {
+      changes = `${columns} = '${values}'`
+    }
+
+    db_comm.run(queries.UPDATE(opts.type, opts.id, changes), function(err) {
+      var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
+      return cb && cb(e);
     });
   });
 }
 
-exports.close = function(type, cb) {
-  remove(type, function(err) {
-    storage_path = null;
-    cb && cb(err);
+storage_fns.all = (opts, cb) => {
+  init(opts.type, null, (err, db) => {
+    if (err) return cb(err);
+
+    // Debería ser db, revisar!
+    db_comm.all(queries.SELECT(opts.type), function(err, rows) {
+      if (err) {
+        if (err.code != 'ENOENT')
+          return cb(err, null);
+
+        return cb(err, []);
+      }
+      cb(null, rows);
+    })
   })
 }
 
-exports.erase = function(path, cb) {
+storage_fns.query = (opts, cb) => {
+  init(opts.type, null, (err) => {
+    let column = opts.column,
+        data = opts.data,
+        query = `${column} = '${data}'`;
+    db_comm.all(queries.QUERY(opts.type, query), function(err, rows) {
+      if (err) {
+        if (err.code != 'ENOENT')
+          return cb(err, null);
+
+        return cb(err, []);
+      }
+      cb(null, rows);
+    })
+  })
+}
+
+storage_fns.clear = (opts, cb) => {
+  db_comm.all(queries.CLEAR(opts.type), (err) => {
+    var e = err && err.code == 'SQLITE_READONLY' ? SQLITE_ACCESS_ERR : err
+    return cb && cb(e);
+  })
+}
+
+// Only for testing purposes
+var erase = function(path, cb) {
   fs.unlink(path, function(err) {
     storage_path = null;
     cb && cb(err);
   })
 }
 
+// Recover data from the old commands.db file and converts it
+var recover_db = function(db, cb) {
+  var get_tables = (cb) => {
+    db.all(queries.TABLES(), (err, data) => {
+      let tables = data.map(x => x.name);
+      tables = tables.filter((x) => { return !x.includes('autoindex') })
+      return cb(tables);
+    });
+  }
+
+  get_tables((tables) => {
+    if (tables.length == 0)
+      return cb();
+
+    let array = [];
+    let data = {};
+
+    tables.forEach((table) => {
+      array.push((callback) => {
+        // Save data from all the tables (except keys)
+        db.all(queries.SELECT(table), (err, rows) => {
+          data[table] = rows;
+          callback();
+        });
+      });
+    })
+
+    async.series(array, (err) => {
+      // If there's only one column it has the old format
+      if (Object.keys(Object.values(data.keys)[0]).length > 1) {
+        return cb();
+      }
+
+      let array2 = [];
+      db.all(queries.DROP('keys'));
+
+      tables.forEach((table) => {
+        array2.push((callback) => {
+        
+          db.all(queries.DROP(table), (err) => {
+            db.run(queries.CREATE(table), () => {
+              data[table].forEach((row) => {
+                let value = JSON.parse(Buffer.from(row[singular(table)], 'base64').toString()),
+                    db_data = Object.values(value)[0],
+                    id;
+
+                if (table == 'commands') {
+                  if (db_data.options && db_data.options.messageID) id = db_data.options.messageID;
+                  else id = uuidv4();
+                }
+                if (table == 'geofences') id = Object.values(value)[0].id;
+                if (table == 'files')     id = Object.keys(value)[0].split('-')[1];
+                if (table == 'triggers')  id = Object.values(value)[0].id;  //validar
+                if (table == 'versions')  id = Object.keys(value)[0].split('-')[1];
+
+                exports.do('set', {type: table, id: id, data: db_data})
+              });
+            });
+          });
+          setTimeout(() => {callback()}, 300);
+        });
+
+      });
+
+      async.series(array2, (err) => {
+        return cb();
+      })
+    });
+  })
+}
+
 exports.init = init;
+exports.erase = erase;
+exports.recover_db = recover_db;
+
+module.exports.storage_fns = storage_fns;
