+++ lib/package.js
@@ -6,7 +6,7 @@ var fs              = require('fs'),
     cp              = require('child_process'),
     whenever        = require('whenever'),
     remove          = require('remover'),
-    storage         = require('./agent/utils/commands_storage'),
+    storage         = require('./agent/utils/storage'),
     is_greater_than = require('./agent/helpers').is_greater_than,
     paths           = require('./system/paths'),
     os_name         = process.platform.replace('win32', 'windows').replace('darwin', 'mac'),
