+++ test/lib/agent/commands.js
@@ -8,8 +8,7 @@ var path      = require('path'),
     root_path = path.resolve(__dirname, '..', '..', '..'),
     lib_path  = path.join(root_path, 'lib'),
     api_path  = join(lib_path, 'agent', 'plugins', 'control-panel', 'api'),
-    storage   = helpers.load('utils/commands_storage'),
-    //storage   = helpers.load('utils/storage'),
+    storage   = helpers.load('utils/storage'),
     keys      = require(join(api_path, 'keys')),
     devices   = require(join(api_path, 'devices')),
     request   = require(join(api_path, 'request')),
